refactor: extract client component
