+++ client/src/components/dashboard/machine/machine-table.tsx
@@ -21,6 +21,7 @@ import { removeMonitoringPointsByMachineId } from '@/store/features/monitoringPo
 import { AppDispatch } from '@/store';
 import { useDispatch } from 'react-redux';
 import { useSnackbar } from '@/providers/SnackProvider';
+import { EditMachineTypeDialogue } from './edit-machine-type-dialogue';
 
 export interface Machine {
   id: string;
@@ -184,32 +185,13 @@ export function MachineTable({ paginatedMachines }: { paginatedMachines: Machine
       ))}
 
       {/* Confirmation Dialog for Machine Type Change */}
-      <Dialog open={confirmDialogOpen} onClose={handleCancelTypeChange}>
-        <DialogTitle>Confirm Machine Type Change</DialogTitle>
-        <DialogContent>
-          <Typography variant="body1" sx={{ mb: 2 }}>
-            Are you sure you want to change the machine type for "{pendingTypeChange?.machineName}" 
-            to {pendingTypeChange?.newType}?
-          </Typography>
-          <Typography variant="body2" color="warning.main" sx={{ fontWeight: 'bold' }}>
-            ⚠️ Warning: This action will delete all monitoring points associated with this machine, 
-            as different machine types have specific sensor types available.
-          </Typography>
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={handleCancelTypeChange} disabled={updatingTypeId !== null}>
-            Cancel
-          </Button>
-          <Button 
-            onClick={handleConfirmTypeChange} 
-            variant="contained" 
-            color="warning"
-            disabled={updatingTypeId !== null}
-          >
-            {updatingTypeId !== null ? 'Updating...' : 'Confirm Change'}
-          </Button>
-        </DialogActions>
-      </Dialog>
+      <EditMachineTypeDialogue
+        confirmDialogOpen={confirmDialogOpen}
+        handleCancelTypeChange={handleCancelTypeChange}
+        handleConfirmTypeChange={handleConfirmTypeChange}
+        updatingTypeId={updatingTypeId}
+        pendingTypeChange={pendingTypeChange}
+      />
     </>
   );
 }
